In Scandir, remove code to check if the key is a directory.
In some cases there is no directory exist as a key, but still there are files in that directory.
This means only if the directory is explicity exist scandir will work in it.
The check for file was removed this means file will also be able to search any file that start with the same name. menaning shared prefix.